Align props with ResourceSectionProps
