+++ src/blocks/article-blocks/ResourceSection/Books/BooksSection.tsx
@@ -1,10 +1,11 @@
-import { Media, ResourceSection } from '@/payload-types'
+import { Media } from '@/payload-types'
 import React from 'react'
 import { Button } from '@/components/ui/button'
 import clsx from 'clsx'
 import { ImagePlaceholder } from '@/features/shared/components/ImagePlaceholder'
 import { RenderMedia } from '@/features/shared/components/RenderMedia'
 import TitleBar from '../../TitleBar'
+import { ResourceWithSlug } from '../types'
 
 type BooksType = {
   image?: Media
@@ -56,7 +57,7 @@ const books: BooksType[] = [
 const buttonStyles =
   'size-14 rounded-full border-4 border-white bg-black text-xl text-white hover:bg-black hover:text-secondary-blue'
 
-export function BooksSection(props: ResourceSection) {
+export function BooksSection(props: ResourceWithSlug) {
   const { title } = props
 
   return (
